Add cast for sortOrder

+++ server/src/main/java/org/eclipse/openvsx/search/SearchService.java
@@ -205,7 +205,7 @@ public class SearchService {
         }
 
         if (sortOrder instanceof String) {
-            sortOrder = Integer.parseInt(sortOrder);
+            (int) sortOrder = Integer.parseInt(sortOrder);
         } else {
             logger.debug(sortOrder.getClass().getName());
         }
